+++ mne/io/tests/test_proc_history.py
@@ -1,3 +1,7 @@
+# Authors: Denis A. Engemann <denis.engemann@gmail.com>
+#          Eric Larson <larson.eric.d@gmail.com>
+# License: Simplified BSD
+
 import numpy as np
 import os.path as op
 from mne import io
@@ -12,7 +16,7 @@ raw_fname = op.join(base_dir, 'test_chpi_raw_sss.fif')
 def test_maxfilter_io():
     """test maxfilter io"""
     raw = io.Raw(raw_fname)
-    mf = raw.info['proc_records'][1]['max_info']
+    mf = raw.info['proc_history'][1]['max_info']
 
     assert_true(mf['sss_info']['frame'], FIFF.FIFFV_COORD_HEAD)
     # based on manual 2.0, rev. 5.0 page 23
@@ -37,7 +41,7 @@ def test_maxfilter_io():
 def test_maxfilter_get_rank():
     """test maxfilter rank lookup"""
     raw = io.Raw(raw_fname)
-    mf = raw.info['proc_records'][0]['max_info']
+    mf = raw.info['proc_history'][0]['max_info']
     rank1 = mf['sss_info']['nfree']
     rank2 = _get_sss_rank(mf)
     assert_equal(rank1, rank2)