+++ mne/tests/test_cov.py
@@ -287,8 +287,8 @@ def test_rank():
                 n_projs_eeg = 0
 
             # check sss
-            if 'proc_records' in this_very_info:
-                mf = this_very_info['proc_records'][0]['max_info']
+            if 'proc_history' in this_very_info:
+                mf = this_very_info['proc_history'][0]['max_info']
                 n_free = _get_sss_rank(mf)
                 if 'mag' not in ch_types and 'grad' not in ch_types:
                     n_free = 0